+++ app/src/src_nonfreenet/org/breezyweather/sources/SourceManager.kt
@@ -102,6 +102,7 @@ import org.breezyweather.sources.openweather.OpenWeatherService
 import org.breezyweather.sources.pagasa.PagasaService
 import org.breezyweather.sources.pirateweather.PirateWeatherService
 import org.breezyweather.sources.recosante.RecosanteService
+import org.breezyweather.sources.rnsa.RnsaService
 import org.breezyweather.sources.smg.SmgService
 import org.breezyweather.sources.smhi.SmhiService
 import org.breezyweather.sources.wmosevereweather.WmoSevereWeatherService
@@ -172,6 +173,7 @@ class SourceManager @Inject constructor(
     pagasaService: PagasaService,
     pirateWeatherService: PirateWeatherService,
     recosanteService: RecosanteService,
+    rnsaService: RnsaService,
     smaScService: SmaScService,
     smaSuService: SmaSuService,
     smgService: SmgService,
@@ -261,6 +263,7 @@ class SourceManager @Inject constructor(
         nwsService,
         pagasaService,
         recosanteService,
+        rnsaService,
         smaScService,
         smaSuService,
         smgService,